+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Bug fixes:
 * [OLMIS-8192](https://openlmis.atlassian.net/browse/OLMIS-8192): Requested quantity explanation column adjusts width with input.
 * [OLMIS-8191](https://openlmis.atlassian.net/browse/OLMIS-8191): Remarks column editable and size adjustable dynamically.
 * [OLMIS-8154](https://openlmis.atlassian.net/browse/OLMIS-8154): Explanation no longer required when requested equals calculated.
+* [OLMIS-8128](https://openlmis.atlassian.net/browse/OLMIS-8128): TB Monthly patients tab no longer shows validation highlights after a page refresh
 
 7.0.16 / 2026-02-05
 =================